requirements.txt
asttokens==2.4.1
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
wcwidth==0.2.13
